requirements.txt
streamlit==1.24.0
moviepy==1.0.3
pydub==0.25.1
google-cloud-speech==2.16.1
google-cloud-texttospeech==2.14.1
openai==0.27.0
requests==2.31.0
Pillow==9.0.0
